가게 리스트 거리순으로 출력

+++ src/main/java/com/depromeet/team5/dto/StoreDto.java
@@ -9,9 +9,9 @@ import java.util.List;
 
 @Data
 public class StoreDto {
-    private Long latitude;
+    private Float latitude;
 
-    private Long longitude;
+    private Float longitude;
 
     private String storeName;
